refactor(proj-gen): using generator objects instead of generator functions for entry and router
