+++ packages/teleport-project-generator-react-basic/src/index.ts
@@ -1,13 +1,27 @@
-import {
-  createRouterIndexFile,
-  createHtmlEntryFile,
-  createComponentGenerator,
-} from './component-generators'
-
 import { createProjectGenerator } from '@teleporthq/teleport-project-generator'
+import { createReactComponentGenerator } from '@teleporthq/teleport-component-generator-react'
+import { createGenerator } from '@teleporthq/teleport-component-generator'
+
+import reactAppRoutingPlugin from '@teleporthq/teleport-plugin-react-app-routing'
+import importStatementsPlugin from '@teleporthq/teleport-plugin-import-statements'
+import prettierJS from '@teleporthq/teleport-postprocessor-prettier-js'
+import prettierHTML from '@teleporthq/teleport-postprocessor-prettier-html'
+
+import { Mapping } from '@teleporthq/teleport-types'
+
+import reactProjectMapping from './react-project-mapping.json'
 
 export const createReactBasicGenerator = () => {
-  const reactComponentGenerator = createComponentGenerator()
+  const reactComponentGenerator = createReactComponentGenerator('CSSModules')
+  reactComponentGenerator.addMapping(reactProjectMapping as Mapping)
+
+  const routingComponentGenerator = createGenerator()
+  routingComponentGenerator.addPlugin(reactAppRoutingPlugin)
+  routingComponentGenerator.addPlugin(importStatementsPlugin)
+  routingComponentGenerator.addPostProcessor(prettierJS)
+
+  const htmlFileGenerator = createGenerator()
+  htmlFileGenerator.addPostProcessor(prettierHTML)
 
   const generator = createProjectGenerator({
     components: {
@@ -19,12 +33,14 @@ export const createReactBasicGenerator = () => {
       path: ['src', 'views'],
     },
     router: {
-      generatorFunction: createRouterIndexFile,
+      generator: routingComponentGenerator,
       path: ['src'],
+      fileName: 'index',
     },
     entry: {
-      generatorFunction: createHtmlEntryFile,
+      generator: htmlFileGenerator,
       path: ['src'],
+      fileName: 'index',
     },
     static: {
       prefix: '/static',
