+++ packages/teleport-project-generator-react-next/src/index.ts
@@ -1,8 +1,20 @@
-import { createReactGenerator, createDocumentFile } from './component-generators'
 import { createProjectGenerator } from '@teleporthq/teleport-project-generator'
+import { createGenerator } from '@teleporthq/teleport-component-generator'
+import { createReactComponentGenerator } from '@teleporthq/teleport-component-generator-react'
+
+import prettierJS from '@teleporthq/teleport-postprocessor-prettier-js'
+
+import { Mapping } from '@teleporthq/teleport-types'
+
+import { createDocumentFileChunks } from './utils'
+import nextMapping from './next-mapping.json'
 
 export const createReactNextGenerator = () => {
-  const reactComponentGenerator = createReactGenerator()
+  const reactComponentGenerator = createReactComponentGenerator('StyledJSX')
+  reactComponentGenerator.addMapping(nextMapping as Mapping)
+
+  const documentFileGenerator = createGenerator()
+  documentFileGenerator.addPostProcessor(prettierJS)
 
   const generator = createProjectGenerator({
     components: {
@@ -18,8 +30,10 @@ export const createReactNextGenerator = () => {
       },
     },
     entry: {
-      generatorFunction: createDocumentFile,
+      generator: documentFileGenerator,
       path: ['pages'],
+      fileName: '_document',
+      chunkGenerationFunction: createDocumentFileChunks,
     },
     static: {
       prefix: '/static',
