+++ packages/teleport-project-generator-react-next/src/utils.ts
@@ -8,10 +8,11 @@ import { generateASTDefinitionForJSXTag } from '@teleporthq/teleport-shared/lib/
 import * as types from '@babel/types'
 
 import { prefixPlaygroundAssetsURL } from '@teleporthq/teleport-shared/lib/utils/uidl-utils'
+import { FILE_TYPE } from '@teleporthq/teleport-shared/lib/constants'
 import { EntryFileOptions } from '@teleporthq/teleport-project-generator/lib/types'
-import { ProjectUIDL } from '@teleporthq/teleport-types'
+import { ProjectUIDL, ChunkDefinition } from '@teleporthq/teleport-types'
 
-export const createDocumentComponentAST = (uidl: ProjectUIDL, options: EntryFileOptions) => {
+export const createDocumentFileChunks = (uidl: ProjectUIDL, options: EntryFileOptions) => {
   const { settings, meta, assets, manifest } = uidl.globals
 
   const htmlNode = generateASTDefinitionForJSXTag('html')
@@ -118,7 +119,20 @@ export const createDocumentComponentAST = (uidl: ProjectUIDL, options: EntryFile
 
   // Create AST representation of the class CustomDocument extends Document
   // https://github.com/zeit/next.js#custom-document
-  return createDocumentWrapperAST(htmlNode)
+  const fileAST = createDocumentWrapperAST(htmlNode)
+
+  const chunks: Record<string, ChunkDefinition[]> = {
+    [FILE_TYPE.JS]: [
+      {
+        name: 'document',
+        type: 'js',
+        content: fileAST,
+        linkAfter: [],
+      },
+    ],
+  }
+
+  return chunks
 }
 
 const createDocumentWrapperAST = (htmlNode, t = types) => {
