+++ packages/teleport-project-generator-vue-nuxt/src/index.ts
@@ -1,8 +1,18 @@
-import { createVueGenerator, createHtmlEntryFile } from './component-generators'
 import { createProjectGenerator } from '@teleporthq/teleport-project-generator'
+import { createVueComponentGenerator } from '@teleporthq/teleport-component-generator-vue'
+import { createGenerator } from '@teleporthq/teleport-component-generator'
+
+import prettierHTML from '@teleporthq/teleport-postprocessor-prettier-html'
+
+import { Mapping } from '@teleporthq/teleport-types'
+
+import nuxtMapping from './nuxt-mapping.json'
 
 export const createVueNuxtGenerator = () => {
-  const vueComponentGenerator = createVueGenerator()
+  const vueComponentGenerator = createVueComponentGenerator(nuxtMapping as Mapping)
+
+  const htmlFileGenerator = createGenerator()
+  htmlFileGenerator.addPostProcessor(prettierHTML)
 
   const generator = createProjectGenerator({
     components: {
@@ -18,7 +28,9 @@ export const createVueNuxtGenerator = () => {
       },
     },
     entry: {
-      generatorFunction: createHtmlEntryFile,
+      generator: htmlFileGenerator,
+      appRootOverride: '{{APP_ROOT}}',
+      fileName: 'app',
       path: [],
     },
     static: {
