+++ packages/teleport-project-generator/src/utils.ts
@@ -1,204 +1,19 @@
 import {
-  addAttributeToNode,
-  addChildNode,
-  addTextNode,
-  addBooleanAttributeToNode,
-} from '@teleporthq/teleport-shared/lib/utils/html-utils'
-
-import {
-  prefixPlaygroundAssetsURL,
   cloneObject,
   traverseElements,
   getComponentFileName,
   getComponentPath,
 } from '@teleporthq/teleport-shared/lib/utils/uidl-utils'
 
-import { slugify } from '@teleporthq/teleport-shared/lib/utils/string-utils'
-import { createHTMLNode } from '@teleporthq/teleport-shared/lib/builders/html-builders'
-
 import {
   GeneratedFile,
   GeneratedFolder,
-  HastNode,
   ProjectUIDL,
   UIDLElement,
   ComponentUIDL,
-  WebManifest,
 } from '@teleporthq/teleport-types'
 
-import { FILE_TYPE } from '@teleporthq/teleport-shared/lib/constants'
-import { DEFAULT_PACKAGE_JSON } from './constants'
-import { EntryFileOptions, PackageJSON, ProjectStrategy } from './types'
-
-export const createHtmlIndexFile = (uidl: ProjectUIDL, options: EntryFileOptions): HastNode => {
-  const { assetsPrefix = '', appRootOverride } = options
-  const { settings, meta, assets, manifest } = uidl.globals
-
-  const htmlNode = createHTMLNode('html')
-  const headNode = createHTMLNode('head')
-  const bodyNode = createHTMLNode('body')
-
-  addChildNode(htmlNode, headNode)
-  addChildNode(htmlNode, bodyNode)
-
-  // Vue and React use a standard <div id="app"/> in the body tag.
-  // Nuxt has an internal templating so requires an override
-  if (appRootOverride) {
-    addTextNode(bodyNode, appRootOverride)
-  } else {
-    const appRootNode = createHTMLNode('div')
-    addAttributeToNode(appRootNode, 'id', 'app')
-    addChildNode(bodyNode, appRootNode)
-  }
-
-  if (settings.language) {
-    addAttributeToNode(htmlNode, 'lang', settings.language)
-  }
-
-  if (settings.title) {
-    const titleTag = createHTMLNode('title')
-    addTextNode(titleTag, settings.title)
-    addChildNode(headNode, titleTag)
-  }
-
-  if (manifest) {
-    const linkTag = createHTMLNode('link') // , { selfClosing: true })
-    addAttributeToNode(linkTag, 'rel', 'manifest')
-    addAttributeToNode(linkTag, 'href', '/static/manifest.json')
-    addChildNode(headNode, linkTag)
-  }
-
-  meta.forEach((metaItem) => {
-    const metaTag = createHTMLNode('meta') // , { selfClosing: true })
-    Object.keys(metaItem).forEach((key) => {
-      const prefixedURL = prefixPlaygroundAssetsURL(assetsPrefix, metaItem[key])
-      addAttributeToNode(metaTag, key, prefixedURL)
-    })
-    addChildNode(headNode, metaTag)
-  })
-
-  assets.forEach((asset) => {
-    const assetPath = prefixPlaygroundAssetsURL(assetsPrefix, asset.path)
-
-    // link stylesheet (external css, font)
-    if ((asset.type === 'style' || asset.type === 'font') && assetPath) {
-      const linkTag = createHTMLNode('link') // , { selfClosing: true })
-      addAttributeToNode(linkTag, 'rel', 'stylesheet')
-      addAttributeToNode(linkTag, 'href', assetPath)
-      addChildNode(headNode, linkTag)
-    }
-
-    // inline style
-    if (asset.type === 'style' && asset.content) {
-      const styleTag = createHTMLNode('style')
-      addTextNode(styleTag, asset.content)
-      addChildNode(headNode, styleTag)
-    }
-
-    // script (external or inline)
-    if (asset.type === 'script') {
-      const scriptInBody = (asset.meta && asset.meta.target === 'body') || false
-      const scriptTag = createHTMLNode('script')
-      // addTextNode(scriptTag, ' ') // To ensure tag is not automatically self-closing, which causes problems in the <head>
-      addAttributeToNode(scriptTag, 'type', 'text/javascript')
-      if (assetPath) {
-        addAttributeToNode(scriptTag, 'src', assetPath)
-        if (asset.meta && asset.meta.defer) {
-          addBooleanAttributeToNode(scriptTag, 'defer')
-        }
-        if (asset.meta && asset.meta.async) {
-          addBooleanAttributeToNode(scriptTag, 'async')
-        }
-      } else if (asset.content) {
-        addTextNode(scriptTag, asset.content)
-      }
-      if (scriptInBody) {
-        addChildNode(bodyNode, scriptTag)
-      } else {
-        addChildNode(headNode, scriptTag)
-      }
-    }
-
-    // icon
-    if (asset.type === 'icon' && assetPath) {
-      const iconTag = createHTMLNode('link') // , { selfClosing: true })
-      addAttributeToNode(iconTag, 'rel', 'shortcut icon')
-      addAttributeToNode(iconTag, 'href', assetPath)
-      if (typeof asset.meta === 'object') {
-        const assetMeta = asset.meta
-        Object.keys(assetMeta).forEach((metaKey) => {
-          addAttributeToNode(iconTag, metaKey, assetMeta[metaKey])
-        })
-      }
-      addChildNode(headNode, iconTag)
-    }
-  })
-
-  return htmlNode
-}
-
-// Creates a manifest json file with the UIDL having priority over the default values
-export const createManifestJSONFile = (uidl: ProjectUIDL, assetsPrefix?: string): GeneratedFile => {
-  const manifest = uidl.globals.manifest
-  const projectName = uidl.name
-  const defaultManifest: WebManifest = {
-    short_name: projectName,
-    name: projectName,
-    display: 'standalone',
-    start_url: '/',
-  }
-
-  const icons = manifest.icons.map((icon) => {
-    const src = prefixPlaygroundAssetsURL(assetsPrefix || '', icon.src)
-    return { ...icon, src }
-  })
-
-  const content = {
-    ...defaultManifest,
-    ...manifest,
-    ...{ icons },
-  }
-
-  return {
-    name: 'manifest',
-    fileType: FILE_TYPE.JSON,
-    content: JSON.stringify(content, null, 2),
-  }
-}
-
-export const handlePackageJSON = (
-  template: GeneratedFolder,
-  uidl: ProjectUIDL,
-  dependencies: Record<string, string>
-) => {
-  const inputPackageJSONFile = template.files.find(
-    (file) => file.name === 'package' && file.fileType === FILE_TYPE.JSON
-  )
-
-  if (inputPackageJSONFile) {
-    const packageJSONContent = JSON.parse(inputPackageJSONFile.content) as PackageJSON
-
-    packageJSONContent.name = slugify(uidl.name)
-    packageJSONContent.dependencies = {
-      ...packageJSONContent.dependencies,
-      ...dependencies,
-    }
-
-    inputPackageJSONFile.content = JSON.stringify(packageJSONContent, null, 2)
-  } else {
-    const content: PackageJSON = {
-      ...DEFAULT_PACKAGE_JSON,
-      name: slugify(uidl.name),
-      dependencies,
-    }
-
-    template.files.push({
-      name: 'package',
-      fileType: FILE_TYPE.JSON,
-      content: JSON.stringify(content, null, 2),
-    })
-  }
-}
+import { ProjectStrategy } from './types'
 
 export const resolveLocalDependencies = (input: ProjectUIDL, strategy: ProjectStrategy) => {
   const result = cloneObject(input)
